services/stream/requirements.txt
faust-streaming==0.11.2
orjson==3.10.7
psycopg[binary]==3.2.2
python-dotenv==1.0.1
